update setup.sh to prompt the user for vars

+++ setup.sh
@@ -1,12 +1,44 @@
 #!/bin/bash
 
-# TODO: vars! (arguments/prompt user!)
-PROJECT_NAME=coolProject
-STD_VERSION=20
-APPS=(app example-one repl)
+# Function to prompt the user for an array input
+function prompt_for_array {
+  local prompt_message=$1
+  local -n array_ref=$2  # Use nameref to pass the array by reference
 
-# remove the generated directory if it exists:
-rm -rf generated/
+  echo "$prompt_message (Separate elements with spaces):"
+  read -a array_ref  # Read input into the array
+}
+
+# Function to prompt the user for a single variable input
+function prompt_for_input {
+  local prompt_message=$1
+  local -n var_ref=$2  # Use nameref to pass the variable by reference
+
+  echo "$prompt_message:"
+  read var_ref  # Read input into the variable
+}
+
+PROJECT_NAME=defaultProject
+STD_VERSION=20
+APPS=(app)
+FORCE=true
+
+prompt_for_input "Enter the Name for your project (no spaces!)" PROJECT_NAME
+prompt_for_array "Enter the list of apps to generate" APPS
+prompt_for_input "Enter the standard version (e.g. 17 for c++17)" STD_VERSION
+
+if [ -d generated ]; then
+  echo ./generated/ already exists
+  if $FORCE; then
+    echo removing existing ./generated folder
+    # remove the generated directory if it exists:
+    rm -rf generated/
+  else
+    exit 1
+  fi
+
+fi
+
 # create the generated directory
 mkdir generated
 
@@ -17,7 +49,7 @@ find . -maxdepth 1 -mindepth 1 ! -name generated -exec cp -r {} generated \;
 cd generated
 
 # remove the unwanted files:
-rm -rf setup.sh README.md .git
+rm -Rf setup.sh README.md .git/
 
 # replace the vars with the ones provided by the user:
